refactor(#3479): extract tmux_watcher supervisor-relay + session-bound ACK cluster to submodule
Pure, behavior-preserving move of the supervisor relay-forward + session-bound
terminal-ACK cluster out of the frozen giant src/services/discord/tmux_watcher.rs
(8223 baseline) into two cohesive sibling child modules. Zero logic change: the
moved fn/struct/enum bodies are byte-identical apart from rustfmt reflowing two
signatures and the added pub(super) visibility markers.

Split into two files only to keep each within the tmux_watcher/** 700-line
namespace cap:
- supervisor_relay.rs (393 prod LoC): the forward half — the shared
  SessionBoundRelayAckTarget / SupervisorRelayForward types,
  carry_session_bound_ack_for_turn, the chunk/terminal forward fns + the
  multi-turn split helper + the inner forward, plus watcher_terminal_commit_fence
  / terminal_event_consumed_offset.
- session_bound_ack.rs (404 prod LoC): the ACK half — SessionBoundRelayAckOutcome
  + the DeliveryOutcome fold, the per-sequence ACK snapshot resolvers, the
  watcher-direct-send gate, WatcherTerminalResendAction + the in-flight-sink-marker
  gate + the committed-offset reconciliation, the RelaySlotGuard emission-slot RAII
  guard, and the ACK delivery wait.

The 13 associated unit tests move verbatim into sibling supervisor_relay_tests.rs
(1) and session_bound_ack_tests.rs (12, incl. the inflight_sink_marker_gate nested
module) so production LoC drops out of the frozen file while the tests still run.

The cluster has ZERO coupling to shared./http./InflightTurnState (verified before
moving each item). The root keeps discard_watcher_pending_buffer_after_suppressed_turn
and Utf8ChunkDecoder/DecodedUtf8Chunk, and re-imports the moved items via
`use self::{supervisor_relay,session_bound_ack}::*;` so all call sites stay intact.

tmux_watcher.rs: 8223 -> 7485 production LoC (-738). Giant ratchet baseline lowered
to 7485; both child modules registered. change-surfaces.md + the auto-generated
inventory (ARCHITECTURE.md) updated. audit ratchet rc=0, ci-script-checks rc=0,
build clean (legacy-sqlite cfg warns only), clippy clean, all tmux_watcher tests
(180) + 13 moved tests green.
