+++ ARCHITECTURE.md
@@ -500,7 +500,11 @@ src/
 │   │   │   ├── panel_decisions.rs
 │   │   │   ├── placeholder_reclaim.rs
 │   │   │   ├── prompt_observe.rs
+│   │   │   ├── session_bound_ack.rs
+│   │   │   ├── session_bound_ack_tests.rs
 │   │   │   ├── single_message_footer.rs
+│   │   │   ├── supervisor_relay.rs
+│   │   │   ├── supervisor_relay_tests.rs
 │   │   │   ├── terminal_send.rs
 │   │   │   ├── turn_identity.rs
 │   │   │   └── turn_identity_tests.rs
