+++ scripts/audit_maintainability_giant_baseline.toml
@@ -45,7 +45,17 @@
 # #3351: +101 for placeholder-reclaim wiring in the relay loop (tick/terminal/
 # fresh-idle reclaim sites + r21 durable-record drops); the helper bodies live
 # in tmux_watcher/placeholder_reclaim.rs. tmux_watcher S2 re-lowers this root.
-"src/services/discord/tmux_watcher.rs" = 8223
+# #3479 Phase-1 rank-1: lowered 8223 -> 7485 by extracting the supervisor
+# relay-forward + session-bound terminal-ACK cluster verbatim (pure move, zero
+# logic change) into two cohesive tmux_watcher/ child modules — supervisor_relay.rs
+# (forward half + shared SessionBoundRelayAckTarget/SupervisorRelayForward types +
+# watcher_terminal_commit_fence/terminal_event_consumed_offset) and
+# session_bound_ack.rs (SessionBoundRelayAckOutcome fold, per-sequence ACK snapshot
+# resolvers, watcher-direct-send gate, WatcherTerminalResendAction + in-flight-sink
+# marker gate, RelaySlotGuard, ACK delivery wait); moved unit tests live in sibling
+# supervisor_relay_tests.rs / session_bound_ack_tests.rs. Split into two child files
+# only to keep each within the tmux_watcher/** 700-line namespace cap.
+"src/services/discord/tmux_watcher.rs" = 7485
 # #3038 tmux_watcher S1 child modules, frozen at landed production LoC.
 "src/services/discord/tmux_watcher/commit_decisions.rs" = 140
 "src/services/discord/tmux_watcher/completion_gate.rs" = 275
@@ -54,6 +64,12 @@
 "src/services/discord/tmux_watcher/placeholder_reclaim.rs" = 106
 "src/services/discord/tmux_watcher/prompt_observe.rs" = 109
 "src/services/discord/tmux_watcher/turn_identity.rs" = 327
+# #3479 Phase-1 rank-1 child modules (supervisor relay-forward + session-bound
+# terminal-ACK cluster), frozen at landed production LoC. Both sit under the
+# tmux_watcher/** 700-line namespace cap; their moved unit tests live in sibling
+# *_tests.rs files (excluded from the production splitter).
+"src/services/discord/tmux_watcher/supervisor_relay.rs" = 393
+"src/services/discord/tmux_watcher/session_bound_ack.rs" = 404
 # #3089 A4 watcher short-replace controller cutover sibling, frozen at landed prod LoC.
 # A4 r2 (codex r1 [High]): +118 for the `apply_watcher_short_replace_result` split
 # + the `DeliveredFallback` write-back arm mirroring the legacy fallback cleanup.
