+++ src/services/discord/tmux_watcher.rs
@@ -44,11 +44,26 @@ mod single_message_footer;
 #[path = "tmux_watcher/terminal_send.rs"]
 mod terminal_send;
 
+// #3479 Phase-1 rank-1: the supervisor relay-forward + session-bound terminal ACK
+// cluster extracted to sibling submodules (pure move, zero logic change). Split
+// into two cohesive files only to keep each within the tmux_watcher/** namespace
+// LoC cap: `supervisor_relay` holds the forward half (+ the shared
+// `SessionBoundRelayAckTarget` type), `session_bound_ack` holds the ACK-outcome /
+// terminal-resend / emission-slot-guard half. Items are `pub(super)` there and
+// re-imported here so the watcher loop's call sites stay byte-identical.
+#[path = "tmux_watcher/supervisor_relay.rs"]
+mod supervisor_relay;
+
+#[path = "tmux_watcher/session_bound_ack.rs"]
+mod session_bound_ack;
+
 pub(in crate::services::discord) use self::completion_gate::{
     TuiCompletionGateOutcome, run_tui_completion_gate,
 };
 use self::placeholder_reclaim::*;
+use self::session_bound_ack::*;
 use self::single_message_footer::*;
+use self::supervisor_relay::*;
 
 fn adopt_watcher_terminal_message_ids_from_inflight(
     placeholder_msg_id: &mut Option<serenity::MessageId>,
@@ -324,118 +339,6 @@ fn watcher_session_ready_for_input(
     }
 }
 
-/// E5 (#2412): forward a freshly-read tmux output chunk into the
-/// supervisor-owned [`StreamRelay`] (if one exists for the session). The
-/// supervisor's [`RelayProducerRegistry`] is the bridge — it hands the
-/// production tmux watcher a clonable
-/// [`crate::services::cluster::stream_relay::RelayProducer`] keyed by
-/// `tmux_session_name`. The producer's MPSC absorbs the chunk; the
-/// relay task drains it into the configured [`RelaySink`]. In production
-/// that sink parses provider JSONL and performs Discord terminal delivery
-/// for eligible session-bound inflight shapes; metrics-only fallback
-/// runtimes still count frames via
-/// [`crate::services::cluster::registry_adapter_sink::RegistryAdapterSink`].
-///
-/// `cached_producer` caches a single producer clone to avoid taking the
-/// registry RwLock on every chunk read; it is refreshed from the registry
-/// when the cache is empty or when an attempted send observed a torn-down
-/// relay (`try_send_frame` returned `false`). When the registry has no
-/// producer for this session (flag off, supervisor not running, or this
-/// session simply not in the registry's matched set) the function is a
-/// total no-op and adds no measurable overhead vs the pre-E5 hot path.
-#[derive(Clone)]
-struct SessionBoundRelayAckTarget {
-    metrics: std::sync::Arc<crate::services::cluster::stream_relay::RelayMetrics>,
-    sequence: u64,
-    /// #3041 P1-3 (codex P1-3 R6): the `turn_start_offset` of the turn this ACK
-    /// target belongs to — taken from the terminal frame's commit fence (the ONLY
-    /// frame that yields an ack target). The watcher's per-turn forward carries the
-    /// stored ack forward ONLY when it belongs to the turn currently being
-    /// ACK-waited (same `turn_start_offset`); a stale ack from a FINISHED/DIFFERENT
-    /// turn is reset to `None` so a new turn never inherits a previous turn's
-    /// terminal sequence (no false-Delivered black-hole). `None` means the fence
-    /// carried no `turn_start_offset` (legacy / no pinned identity) — treated as
-    /// "no turn binding", so it is never reused across a turn boundary.
-    turn_start_offset: Option<u64>,
-}
-
-#[derive(Clone)]
-struct SupervisorRelayForward {
-    mirrored: bool,
-    ack_target: Option<SessionBoundRelayAckTarget>,
-    /// #3041 P1-3 (codex P1-3 R7): TRUE when this forward SPLIT a result-bearing
-    /// physical chunk and a NON-EMPTY trailing tail (a LATER turn's bytes) followed
-    /// the just-completed turn's terminal frame. This is the turn-boundary signal:
-    /// after the just-completed turn (A) consumes its own terminal ACK, the watcher
-    /// must RESET the stored ack to `None` so the trailing turn (B) — which is
-    /// processed from the leftover buffer on a later pass, possibly while
-    /// `turn_identity_for_panel` is STILL pinned to A's offset — can NEVER inherit
-    /// A's finished ACK. With no inherited ack B reads `MissingTarget` → §3.2
-    /// reconciliation (committed-offset SendFull-or-Skip) → B is never black-holed.
-    /// Only the SPLIT terminal forward sets this; every other forward leaves it
-    /// `false` (no boundary crossed).
-    trailing_turn_follows: bool,
-}
-
-impl SupervisorRelayForward {
-    fn mirrored_without_ack() -> Self {
-        Self {
-            mirrored: true,
-            ack_target: None,
-            trailing_turn_follows: false,
-        }
-    }
-
-    fn not_mirrored() -> Self {
-        Self {
-            mirrored: false,
-            ack_target: None,
-            trailing_turn_follows: false,
-        }
-    }
-}
-
-/// #3041 P1-3 (codex P1-3 R6): turn-scope the session-bound terminal ACK target so
-/// a NEW turn never inherits a FINISHED turn's stale ack.
-///
-/// The watcher carries `all_data_session_bound_relay_ack` across `'watcher_loop`
-/// passes. A single chunk can hold `result(A) + assistant(B) + result(B)`: A rides
-/// a terminal frame (ack = A.seq) while B finishes inside the discarded split tail.
-/// On the next pass B is processed from the leftover buffer with no fresh frame, so
-/// under the legacy "store only when `Some`" rule the stored ack stays pinned to A
-/// and A's `Delivered` FALSELY satisfies B's ACK → B black-holed.
-///
-/// The ack is keyed by `current_turn_start_offset` (the same
-/// `InflightTurnIdentity.turn_start_offset` the terminal fence stamps — monotonic
-/// per turn — NOT the per-pass `turn_data_start_offset`):
-///   * `fresh` is `Some` → this pass forwarded a terminal frame; adopt it.
-///   * `fresh` is `None` → keep `stored` ONLY when it is the SAME turn
-///     (`stored.turn_start_offset == current_turn_start_offset`, both `Some`); a
-///     cross-turn / unbound `stored` drops to `None`. A `None` target makes
-///     `wait_for_session_bound_relay_delivery_ack` return `MissingTarget` (NOT
-///     `Delivered`), so the watcher falls through to §3.2 reconciliation against
-///     `committed_relay_offset` → re-sent at worst (possible duplicate), NEVER
-///     black-holed.
-fn carry_session_bound_ack_for_turn(
-    stored: Option<SessionBoundRelayAckTarget>,
-    fresh: Option<SessionBoundRelayAckTarget>,
-    current_turn_start_offset: Option<u64>,
-) -> Option<SessionBoundRelayAckTarget> {
-    if let Some(fresh) = fresh {
-        return Some(fresh);
-    }
-    match (stored, current_turn_start_offset) {
-        // Same turn (both bound to the same pinned `turn_start_offset`): an ack set
-        // earlier in THIS turn survives a later non-terminal pass.
-        (Some(ack), Some(current)) if ack.turn_start_offset == Some(current) => Some(ack),
-        // A stale ack from a finished/different turn — or any case where the turn
-        // binding is unknown on either side — is never consulted: reset to `None`
-        // so the new turn reconciles instead of satisfying its ACK against the
-        // previous turn's sequence.
-        _ => None,
-    }
-}
-
 fn discard_watcher_pending_buffer_after_suppressed_turn(
     all_data: &mut String,
     all_data_start_offset: &mut u64,
@@ -519,647 +422,6 @@ impl Utf8ChunkDecoder {
         self.pending.clear();
         self.pending_start_offset = None;
     }
-}
-
-fn forward_chunk_to_supervisor_relay(
-    tmux_session_name: &str,
-    chunk: &str,
-    registry: &std::sync::Arc<
-        crate::services::cluster::relay_producer_registry::RelayProducerRegistry,
-    >,
-    cached_producer: &mut Option<crate::services::cluster::stream_relay::RelayProducer>,
-) -> SupervisorRelayForward {
-    forward_chunk_to_supervisor_relay_inner(
-        tmux_session_name,
-        chunk,
-        registry,
-        cached_producer,
-        None,
-    )
-}
-
-/// #3041 P1-3 (Part a, B1): forward the RESULT-bearing chunk as a TERMINAL frame
-/// carrying the commit fence (`terminal.consumed_end` + the pinned turn identity).
-/// Every non-terminal chunk goes through `forward_chunk_to_supervisor_relay` with
-/// no fence (unchanged behaviour). Only the result-bearing chunk — detected AFTER
-/// `process_watcher_lines` sets `found_result` — uses this so the commit data rides
-/// the exact frame that triggers the sink's terminal delivery (FIFO single-task: a
-/// separate later frame would arrive after the delivery already dispatched).
-fn forward_terminal_chunk_to_supervisor_relay(
-    tmux_session_name: &str,
-    chunk: &str,
-    registry: &std::sync::Arc<
-        crate::services::cluster::relay_producer_registry::RelayProducerRegistry,
-    >,
-    cached_producer: &mut Option<crate::services::cluster::stream_relay::RelayProducer>,
-    terminal: crate::services::cluster::stream_relay::TerminalCommitFence,
-) -> SupervisorRelayForward {
-    forward_chunk_to_supervisor_relay_inner(
-        tmux_session_name,
-        chunk,
-        registry,
-        cached_producer,
-        Some(terminal),
-    )
-}
-
-/// #3041 P1-3 (codex P1-3 issue 1): forward a RESULT-bearing physical chunk that
-/// may ALSO contain a trailing LATER-turn tail. `leftover_len` is the post-parse
-/// `all_data.len()` — the bytes `process_watcher_lines` did NOT consume (the next
-/// turn's bytes). We split `decoded` at that boundary and:
-///   1. forward the just-completed turn's bytes (`terminal_part`) on a TERMINAL
-///      frame carrying THIS turn's commit fence, and
-///   2. forward the trailing later-turn bytes (`tail_part`) on a SEPARATE
-///      NON-terminal frame so they are still mirrored into the sink's parser and
-///      the later turn is never black-holed (it gets its own fence when it
-///      completes on a later pass).
-///
-/// The returned ACK target is the TERMINAL frame's (so the watcher's terminal-ACK
-/// wait correlates to THIS turn's delivery, not the trailing fragment). `mirrored`
-/// is the AND of both forwards. When there is no trailing tail this is exactly the
-/// single terminal forward.
-fn forward_terminal_chunk_with_trailing_to_supervisor_relay(
-    tmux_session_name: &str,
-    decoded: &str,
-    leftover_len: usize,
-    registry: &std::sync::Arc<
-        crate::services::cluster::relay_producer_registry::RelayProducerRegistry,
-    >,
-    cached_producer: &mut Option<crate::services::cluster::stream_relay::RelayProducer>,
-    terminal: crate::services::cluster::stream_relay::TerminalCommitFence,
-) -> SupervisorRelayForward {
-    let (terminal_part, tail_part) =
-        split_decoded_chunk_at_terminal_boundary(decoded, leftover_len);
-    let terminal_forward = forward_terminal_chunk_to_supervisor_relay(
-        tmux_session_name,
-        terminal_part,
-        registry,
-        cached_producer,
-        terminal,
-    );
-    if tail_part.is_empty() {
-        return terminal_forward;
-    }
-    // Forward the later-turn tail as its OWN non-terminal frame (no fence). This
-    // keeps it in the sink's parser stream so the later turn is mirrored; its
-    // terminal fence rides a future result-bearing chunk. We keep the TERMINAL
-    // frame's ack_target (the watcher waits on THIS turn's delivery) and AND the
-    // tail's mirror flag so a failed tail forward still surfaces "not fully
-    // mirrored".
-    //
-    // #3041 P1-3 (codex P1-3 issue 1 R4 — DEFERRED multi-RESULT edge, #3151): a
-    // per-result split that gives turn B its OWN terminal fence is INFEASIBLE in
-    // this pass — a fence requires B's PINNED turn identity (`turn_start_offset` +
-    // `user_msg_id` + `started_at`), but only turn A's identity
-    // (`turn_identity_for_panel`) is loaded here; B's inflight is established on a
-    // LATER watcher loop pass. Any fence we emitted for B's bytes would carry A's
-    // identity and the sink's STRICT identity gate would (correctly) BLOCK it. So B
-    // rides this fence-less tail: it is MIRRORED (no black-hole) and gets its own
-    // real fence when B completes on a later pass. If this tail already contains
-    // B's COMPLETE result, the sink posts B from it; the watcher's later SendFull
-    // may then re-post B → a possible DUPLICATE (never a black-hole). The
-    // ACK-correlation hazard is closed in the sink (`deliver`): a fence-less frame
-    // reports `FrameAccepted`, NEVER a terminal commit, so B's post can never
-    // satisfy turn A's terminal-ACK and the ACK stays bound to A's terminal frame.
-    let tail_forward =
-        forward_chunk_to_supervisor_relay(tmux_session_name, tail_part, registry, cached_producer);
-    SupervisorRelayForward {
-        mirrored: terminal_forward.mirrored && tail_forward.mirrored,
-        ack_target: terminal_forward.ack_target,
-        // #3041 P1-3 (codex P1-3 R7): a NON-EMPTY trailing tail means a LATER turn's
-        // bytes followed THIS turn's terminal frame inside ONE physical chunk — a
-        // turn-boundary signal. The watcher resets the stored ack AFTER this turn
-        // consumes its own terminal ACK, so the trailing turn never inherits this
-        // finished turn's ACK (R7 black-hole close), regardless of whether
-        // `turn_identity_for_panel` has refreshed to the trailing turn yet.
-        trailing_turn_follows: true,
-    }
-}
-
-/// #3041 P1-3 (codex P1-3 issue 1 — multi-turn-chunk black-hole close): split the
-/// freshly-decoded physical chunk at the consumed-terminal boundary so a TERMINAL
-/// frame carries ONLY the just-completed turn's bytes, and the trailing bytes of a
-/// LATER turn ride a SEPARATE (non-terminal) frame.
-///
-/// A single physical read can contain turn A's `result` PLUS turn B's first bytes.
-/// `process_watcher_lines` stops at A's `result` and leaves B's bytes in the
-/// outer-scope `all_data` (the `leftover_len` after the parse). If we forwarded the
-/// WHOLE chunk on A's terminal frame, B's bytes would be consumed by the sink's
-/// parser as part of A's frame; on the NEXT loop pass `decoded.text` can be empty,
-/// so `forward_chunk_to_supervisor_relay_inner` emits NO frame for B → B is never
-/// delivered (black-hole), and the now-stale ACK for A can be reused for B
-/// (mis-commit). Splitting here forwards A's bytes terminal (with A's fence) and
-/// B's trailing bytes as their own non-terminal frame, so B is still mirrored and
-/// — when B completes — gets its OWN terminal frame + fence on a later pass.
-///
-/// The trailing `min(decoded.len(), leftover_len)` bytes of `decoded` are the part
-/// that survived the parse as leftover (B's bytes); the rest is A's terminal
-/// payload. `leftover_len` is the post-parse `all_data.len()` (bytes the parser did
-/// NOT consume). The split index is clamped to a UTF-8 char boundary (defensive —
-/// the leftover always begins on a JSONL `\n` line boundary in practice).
-fn split_decoded_chunk_at_terminal_boundary(decoded: &str, leftover_len: usize) -> (&str, &str) {
-    let trailing = leftover_len.min(decoded.len());
-    let mut split = decoded.len() - trailing;
-    while split < decoded.len() && !decoded.is_char_boundary(split) {
-        // A multibyte scalar straddles the nominal split: keep it whole on the
-        // terminal side rather than panic-slicing mid-scalar. The leftover begins
-        // one boundary later; the sink reorders nothing within a single line, so a
-        // whole extra line on the terminal side is harmless and never drops bytes.
-        split += 1;
-    }
-    decoded.split_at(split)
-}
-
-fn forward_chunk_to_supervisor_relay_inner(
-    tmux_session_name: &str,
-    chunk: &str,
-    registry: &std::sync::Arc<
-        crate::services::cluster::relay_producer_registry::RelayProducerRegistry,
-    >,
-    cached_producer: &mut Option<crate::services::cluster::stream_relay::RelayProducer>,
-    terminal: Option<crate::services::cluster::stream_relay::TerminalCommitFence>,
-) -> SupervisorRelayForward {
-    if chunk.is_empty() {
-        return SupervisorRelayForward::mirrored_without_ack();
-    }
-    if cached_producer.is_none() {
-        *cached_producer = registry.get_producer(tmux_session_name);
-    }
-    let Some(producer) = cached_producer.as_ref() else {
-        return SupervisorRelayForward::not_mirrored();
-    };
-    // The relay treats each `try_send_frame` call as one frame. The caller
-    // decodes only complete UTF-8 prefixes, so a multibyte scalar split across
-    // file reads is forwarded after the next read completes it instead of being
-    // replaced with U+FFFD.
-    let payload = chunk.to_string();
-    // #3041 P1-3 R6: capture the terminal frame's `turn_start_offset` BEFORE the
-    // fence is moved into the send so the resulting ack target can be turn-scoped
-    // (a stored ack is reused across a watcher pass ONLY when it belongs to the
-    // turn now being ACK-waited). A non-terminal frame has no fence → no ack
-    // target is produced (the `outcome.sequence.map` below yields `None`).
-    let ack_turn_start_offset = terminal.as_ref().and_then(|fence| fence.turn_start_offset);
-    let outcome = match terminal {
-        Some(fence) => producer.try_send_terminal_frame_with_sequence(payload, fence),
-        None => producer.try_send_frame_with_sequence(payload),
-    };
-    if !outcome.is_alive() {
-        // Relay was torn down between our registry read and the send —
-        // drop the cache so the next chunk re-resolves. If the supervisor
-        // republishes for the same session name (Updated event), the
-        // next call will hit the new producer.
-        *cached_producer = None;
-        return SupervisorRelayForward::not_mirrored();
-    }
-    SupervisorRelayForward {
-        mirrored: true,
-        ack_target: outcome.sequence.map(|sequence| SessionBoundRelayAckTarget {
-            metrics: producer.metrics().clone(),
-            sequence,
-            turn_start_offset: ack_turn_start_offset,
-        }),
-        // A single forward of one frame never crosses a turn boundary; only the
-        // split helper sets this when it forwards a separate trailing tail.
-        trailing_turn_follows: false,
-    }
-}
-
-/// #3041 P1-5: the watcher's view of the session-bound terminal ACK. The
-/// non-failure arms fold 1:1 onto the cross-actor 3-way `DeliveryOutcome`:
-///   * `Delivered`      ← ring `DeliveryOutcome::Delivered`
-///   * `NotDelivered`   ← ring `DeliveryOutcome::NotDelivered` (the former
-///                        `TerminalSkipped`; a deterministic sink decline)
-///   * the failure/unconfirmed arms (`Unknown`-class) — `RingUnknown` (the ring
-///     recorded an explicit `Unknown`: sink POSTed without confirming),
-///     `Dropped`, `SinkError`, `TimedOut`, `MissingTarget` — ALL collapse to
-///     `DeliveryOutcome::Unknown` for the resend DECISION (see
-///     [`session_bound_ack_delivery_outcome`]). They stay DISTINCT variants here
-///     so the flight-recorder / metrics keep their exact provenance.
-///
-/// §3.2 SAFETY INVARIANT: BOTH `NotDelivered` AND every `Unknown`-class arm route
-/// through `watcher_terminal_resend_action` (committed-offset reconciliation).
-/// There is NO blind skip for `NotDelivered` and NO blind 10s re-send for any
-/// `Unknown`-class arm.
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-enum SessionBoundRelayAckOutcome {
-    Delivered,
-    NotDelivered,
-    RingUnknown,
-    Dropped,
-    SinkError,
-    TimedOut,
-    MissingTarget,
-}
-
-/// #3041 P1-5: collapse the watcher ACK onto the canonical cross-actor 3-way
-/// `DeliveryOutcome` for the resend DECISION. `Delivered` → delivered (no resend);
-/// `NotDelivered` → not-delivered (reconcile); every failure/unconfirmed arm →
-/// `Unknown` (reconcile). The §3.2 reconciliation treats `NotDelivered` and
-/// `Unknown` IDENTICALLY (both consult the committed offset → SendFull-or-Skip),
-/// so this fold is what guarantees neither gets a blind fast-path.
-fn session_bound_ack_delivery_outcome(
-    ack_outcome: SessionBoundRelayAckOutcome,
-) -> crate::services::cluster::stream_relay::DeliveryOutcome {
-    use crate::services::cluster::stream_relay::DeliveryOutcome;
-    match ack_outcome {
-        SessionBoundRelayAckOutcome::Delivered => DeliveryOutcome::Delivered,
-        SessionBoundRelayAckOutcome::NotDelivered => DeliveryOutcome::NotDelivered,
-        SessionBoundRelayAckOutcome::RingUnknown
-        | SessionBoundRelayAckOutcome::Dropped
-        | SessionBoundRelayAckOutcome::SinkError
-        | SessionBoundRelayAckOutcome::TimedOut
-        | SessionBoundRelayAckOutcome::MissingTarget => DeliveryOutcome::Unknown,
-    }
-}
-
-fn sequence_reached(latest: Option<u64>, target: u64) -> bool {
-    latest.is_some_and(|sequence| sequence >= target)
-}
-
-fn session_bound_relay_ack_snapshot_outcome(
-    target: Option<&SessionBoundRelayAckTarget>,
-) -> Option<SessionBoundRelayAckOutcome> {
-    use crate::services::cluster::stream_relay::DeliveryOutcome;
-    let target = target?;
-    // #3041 P1-3 R5 (per-sequence terminal-ACK correlation): resolve the terminal
-    // ACK on THIS watcher's OWN terminal frame (`target.sequence`) EXACT outcome,
-    // NOT the `>=` high-water-mark. When two turns share a physical chunk (turn A
-    // frame seq N, turn B tail seq N+1), B committing bumps the high-water-mark to
-    // N+1; the old `committed >= N` test would then falsely report A as Delivered
-    // even when A's own terminal frame was SKIPPED — black-holing A. Keying the
-    // ACK to A's exact sequence decouples A from B: A reads outcome[N] (its own
-    // result), B reads outcome[N+1]. `None` (not yet resolved / dropped / evicted)
-    // falls through so a dropped/lagging frame keeps waiting → eventually TimedOut
-    // → the watcher reconciles against the committed offset (no false ACK).
-    match target
-        .metrics
-        .terminal_outcome_for_sequence(target.sequence)
-    {
-        Some(DeliveryOutcome::Delivered) => {
-            return Some(SessionBoundRelayAckOutcome::Delivered);
-        }
-        Some(DeliveryOutcome::NotDelivered) => {
-            return Some(SessionBoundRelayAckOutcome::NotDelivered);
-        }
-        // #3041 P1-5: an explicit ring `Unknown` (sink POSTed but could not confirm
-        // the commit) RESOLVES the per-sequence ACK immediately to a `RingUnknown`
-        // — the watcher reconciles against the committed offset NOW instead of
-        // waiting out the 10s ACK timeout. `RingUnknown` folds to
-        // `DeliveryOutcome::Unknown`, which §3.2 treats exactly like `NotDelivered`
-        // (committed-offset SendFull-or-Skip), so this is a faster path to the SAME
-        // safe reconciliation — never a blind re-send.
-        Some(DeliveryOutcome::Unknown) => {
-            return Some(SessionBoundRelayAckOutcome::RingUnknown);
-        }
-        None => {}
-    }
-    // Sink-error / drop remain high-water-mark signals (terminal outcome was never
-    // recorded for this sequence in those paths): they are per-sequence-monotonic
-    // failure markers, not a co-chunked-turn confusion vector.
-    let snapshot = target.metrics.snapshot();
-    if sequence_reached(snapshot.last_sink_error_sequence, target.sequence) {
-        return Some(SessionBoundRelayAckOutcome::SinkError);
-    }
-    if sequence_reached(snapshot.last_dropped_sequence, target.sequence) {
-        return Some(SessionBoundRelayAckOutcome::Dropped);
-    }
-    None
-}
-
-fn session_bound_relay_frame_ack_reached(target: Option<&SessionBoundRelayAckTarget>) -> bool {
-    let Some(target) = target else {
-        return false;
-    };
-    let snapshot = target.metrics.snapshot();
-    sequence_reached(snapshot.last_delivered_sequence, target.sequence)
-}
-
-fn watcher_should_direct_send_after_session_bound_ack(
-    should_direct_send: bool,
-    ack_outcome: SessionBoundRelayAckOutcome,
-    relay_owner_present: bool,
-) -> bool {
-    use crate::services::cluster::stream_relay::DeliveryOutcome;
-    // #3042 (relay-stability P1, OBSOLETE band-aid — removed by #3041 P1-5): #3042
-    // early-`return false`d an ownerless (post-restart restore_inflight gap)
-    // `TimedOut` to blanket-suppress the watcher-direct fallback (rationale: the sink
-    // "may have posted but failed to advance the committed metric" → blind re-send
-    // → 3× duplicate). No longer holds: #3041 P1-3 Part (a)
-    // (`advance_offset_for_confirmed_delegated_terminal`, session_relay_sink.rs ~459)
-    // now couples a CONFIRMED sink POST to advancing `confirmed_end_offset` to the
-    // fenced `end`; a `TimedOut` (not `MissingTarget`) is ONLY produced for a FENCED
-    // forwarded frame (~2038/2053), the same fence the sink advances on, so the
-    // committed offset reflects a confirmed post even ownerless (owner-independent
-    // atomic). The blanket suppression is thus obsolete and HARMFUL — it returned
-    // before §3.2 committed-offset reconciliation, so an ownerless `TimedOut` with
-    // committed < end neither reconciled nor resent (black-hole). Routing through
-    // §3.2 instead: committed >= end → `SkipAlreadyCommitted` (the 3× duplicate is
-    // prevented PRINCIPALLY); committed < end → `SendFull` (black-hole closed). This
-    // completes the P1-5 §3.2 invariant: EVERY non-`Delivered` outcome routes through
-    // committed-offset reconciliation — none blind-skips/resends. (`relay_owner_present`
-    // stays in the signature for the telemetry call site though the gate ignores it.)
-    let _ = relay_owner_present;
-    // #3041 P1-5: decide on the cross-actor 3-way `DeliveryOutcome` instead of the
-    // implicit `ack_outcome != Delivered` bit. `Delivered` → no watcher re-send.
-    // `NotDelivered` AND `Unknown` (every failure/unconfirmed arm) BOTH intend a
-    // re-send here — but that intent is only the PRECONDITION GATE; the actual send
-    // is masked downstream by `watcher_terminal_resend_action` (committed-offset
-    // reconciliation), so neither gets a blind skip (NotDelivered) nor a blind
-    // re-send (Unknown). §3.2 SAFETY INVARIANT.
-    should_direct_send
-        && !matches!(
-            session_bound_ack_delivery_outcome(ack_outcome),
-            DeliveryOutcome::Delivered
-        )
-}
-
-/// #3041 P1-3 (Part b, §3.2): the watcher's terminal re-send DECISION after a
-/// non-`Delivered` session-bound ACK, reconciled against the offset authority
-/// (`committed_relay_offset`) instead of BLINDLY re-sending (removes the 10s
-/// `relay_terminal_ack_timeout` duplicate vector). `committed >= end` → already
-/// delivered (ACK merely lagged) → SKIP; `committed < end` → re-send the FULL
-/// response (no black-hole).
-///
-/// codex BLOCKER 2 (no SendSuffix for the watcher path): the watcher delivers
-/// RESPONSE TEXT sliced by `response_sent_offset` — a DIFFERENT coordinate
-/// system from the JSONL byte `committed`/`start`/`end` — so a suffix slice
-/// cannot be mapped correctly. The sink-delegated terminal delivery is also
-/// ALL-OR-NOTHING (the sink advances to the FULL `end` only after one confirmed
-/// `replace_message_with_outcome`), so the partial-overlap middle case
-/// effectively does not occur and SendFull on `committed < end` is SAFE. Only
-/// the watcher response-text path is restricted to Skip/Full; the idle-relay
-/// path keeps its real JSONL suffix re-read.
-///
-/// #3041 P1-3 issue 4 (DEFERRED, #3151): the 10s ACK wait can elapse while the
-/// sink's POST is still IN FLIGHT → `committed < end` → SendFull → a duplicate
-/// when the in-flight POST later succeeds. Not a regression (the pre-P1-3 path
-/// also re-sent on timeout; SendFull IS the retry, no black-hole); the full fix
-/// is the in-flight sink-delivery marker tracked by #3151.
-#[derive(Debug, PartialEq, Eq, Clone, Copy)]
-enum WatcherTerminalResendAction {
-    /// `committed >= end`: the whole range is already delivered. Do NOT re-send.
-    SkipAlreadyCommitted,
-    /// `committed < end`: the range is not (fully) covered — re-send the full
-    /// response. See the type doc for why no partial-suffix variant exists for
-    /// the watcher response-text path (coordinate mismatch + all-or-nothing sink).
-    ///
-    /// #3041 P1-3 (codex P1-3 issue 4 — DEFERRED, #3151): this arm also fires when
-    /// the sink's POST is still IN FLIGHT at the 10s ACK timeout (committed has not
-    /// advanced yet) → a duplicate once that POST succeeds. No regression (the
-    /// pre-P1-3 path re-sent on timeout too) and no black-hole (SendFull is the
-    /// retry). The remaining slow-sink-in-flight duplicate is closed by the future
-    /// in-flight sink-delivery marker tracked in #3151.
-    SendFull,
-    /// #3151: a sink POST is genuinely IN FLIGHT for this range (the per-channel
-    /// `DeliveryLeaseCell` is `Leased{Sink, fresh}`). The watcher must NOT re-send
-    /// this pass — neither SendFull nor a Skip-log — and let its NEXT terminal pass
-    /// re-evaluate. This is a BOUNDED wait: each pass re-reads the cell, and within
-    /// at most one `DELIVERY_LEASE_DEADLINE_MS` the sink either commits+releases
-    /// (→ committed >= end → Skip) or dies (→ deadline lapses → reclaim + SendFull).
-    /// No busy-loop is introduced — it rides the existing watcher iteration cadence.
-    WaitInFlight,
-}
-
-/// #3151: gate the watcher terminal re-send on the in-flight sink-delivery marker
-/// BEFORE deferring to [`watcher_terminal_resend_action`]. The marker is a
-/// `Leased{Sink, ..}` state on the per-channel `DeliveryLeaseCell`; this reads a
-/// coherent `snapshot` (materialized under the cell's payload mutex) and decides:
-///
-/// - `Leased{Sink}` AND `now_ms < deadline_ms` → [`WatcherTerminalResendAction::WaitInFlight`]
-///   (a sink POST is genuinely in flight — do not re-send this pass).
-/// - `Leased{Sink}` AND `now_ms >= deadline_ms` → RECLAIM (the caller force-clears
-///   the dead sink's marker via `reclaim_if_expired`) then fall through to
-///   `watcher_terminal_resend_action` → `SendFull` (committed < end). No black-hole.
-/// - `Committed{Sink}` (sink committed its terminal decision, not yet released) →
-///   route through the committed-offset reconciliation: `committed >= end` (a real
-///   Delivered commit) → Skip, `committed < end` (a NotDelivered / refused-advance
-///   commit) → SendFull (re-send; no black-hole). #3159 BUG 1: the marker is no
-///   longer blindly treated as delivered.
-/// - ANY non-Sink holder / `Unleased` / committed-covered → behave EXACTLY as today:
-///   defer to `watcher_terminal_resend_action`. The gate ONLY interposes for a
-///   Sink-held lease, so the watcher-direct B2 path is untouched.
-///
-/// Returns `(action, reclaim_expired_sink)`. When `reclaim_expired_sink` is true
-/// the caller MUST call `reclaim_if_expired(now_ms)` on the cell before sending
-/// (the side effect is kept out of this pure decision fn so it stays unit-testable).
-fn watcher_terminal_resend_action_gated(
-    snapshot: &crate::services::discord::LeaseSnapshot,
-    committed: u64,
-    start: u64,
-    end: u64,
-    now_ms: u64,
-) -> (WatcherTerminalResendAction, bool) {
-    use crate::services::discord::{LeaseHolder, LeaseSnapshot};
-    match snapshot {
-        LeaseSnapshot::Leased {
-            holder: LeaseHolder::Sink,
-            deadline_ms,
-            ..
-        } => {
-            if now_ms < *deadline_ms {
-                // Live, in-flight sink POST — wait this pass (bounded by deadline).
-                (WatcherTerminalResendAction::WaitInFlight, false)
-            } else {
-                // Dead/stalled sink — reclaim its marker and re-send (no black-hole).
-                (watcher_terminal_resend_action(committed, start, end), true)
-            }
-        }
-        LeaseSnapshot::Committed {
-            holder: LeaseHolder::Sink,
-            ..
-        } => {
-            // #3159 BUG 1: a Committed{Sink} marker is NO LONGER assumed delivered.
-            // Route through the committed-offset reconciliation; committed >= end
-            // (a real Delivered commit, which advanced the offset BEFORE committing)
-            // → SkipAlreadyCommitted (unchanged), committed < end (a NotDelivered /
-            // refused-advance commit) → SendFull (the range was NOT delivered, so
-            // re-send; no black-hole). This also subsumes the Drop-release fallback
-            // (Unleased + committed < end → SendFull). The committed offset is the
-            // sole delivered-test, so a genuinely-delivered range is never re-sent.
-            (watcher_terminal_resend_action(committed, start, end), false)
-        }
-        // Unleased, or held/committed by a non-Sink holder (Watcher/Bridge): the
-        // #3151 marker does not apply — behave exactly as the pre-#3151 path.
-        _ => (watcher_terminal_resend_action(committed, start, end), false),
-    }
-}
-
-/// Reconcile a watcher terminal re-send against the committed offset authority.
-/// Only ever consulted when the watcher WOULD have re-sent (a non-`Delivered`
-/// ACK and a real body); the caller still applies the existing `relay_owner`
-/// suppression. A zero/inverted range (`end <= start`) yields `SendFull` so the
-/// existing zero-range guards (which never lease/advance) stay in control — the
-/// reconciliation never manufactures a skip for a range it cannot reason about.
-fn watcher_terminal_resend_action(
-    committed: u64,
-    start: u64,
-    end: u64,
-) -> WatcherTerminalResendAction {
-    if end <= start {
-        // Degenerate range: defer to the existing no-range handling downstream.
-        return WatcherTerminalResendAction::SendFull;
-    }
-    if committed >= end {
-        WatcherTerminalResendAction::SkipAlreadyCommitted
-    } else {
-        // committed < end (incl. the partial `start < committed < end` case which
-        // the all-or-nothing sink delegation does not actually produce): re-send
-        // the FULL response. No black-hole; no mis-offset suffix (codex BLOCKER 2).
-        WatcherTerminalResendAction::SendFull
-    }
-}
-
-fn watcher_terminal_response_for_direct_send<'a>(
-    full_response: &'a str,
-    response_sent_offset: usize,
-    session_bound_fallback_uses_full_body: bool,
-) -> &'a str {
-    if session_bound_fallback_uses_full_body {
-        return full_response;
-    }
-    full_response.get(response_sent_offset..).unwrap_or("")
-}
-
-fn watcher_should_send_ordered_new_chunks_for_terminal_fallback(
-    session_bound_fallback_uses_full_body: bool,
-    relay_text: &str,
-) -> bool {
-    session_bound_fallback_uses_full_body
-        && relay_text.len() > crate::services::discord::DISCORD_MSG_LIMIT
-}
-
-/// #2840 (relay-stability P1): RAII guard for the cross-watcher emission slot
-/// (`relay_coord.relay_slot`, an `Arc<AtomicU64>`: 0 = free, non-zero = a
-/// watcher is mid-emission with that start offset). The slot is shared across
-/// every watcher instance for a channel/session, so if the holding watcher
-/// early-returns, hits a `?`, panics, or is task-aborted between CAS-acquire
-/// and the manual `store(0)`, the slot stays non-zero forever and every
-/// replacement watcher's relay is skipped — a permanent channel wedge until
-/// process restart.
-///
-/// The guard releases the slot on Drop so ANY exit path frees it. The two
-/// intended in-loop release points still call `release()` explicitly to
-/// preserve their exact timing (site 1 releases *before* a 500ms backoff sleep,
-/// so scope-end Drop alone would hold the slot across that sleep); the
-/// idempotent `released` flag makes the trailing Drop a no-op after an explicit
-/// release.
-struct RelaySlotGuard {
-    slot: std::sync::Arc<std::sync::atomic::AtomicU64>,
-    released: bool,
-}
-
-impl RelaySlotGuard {
-    fn new(slot: std::sync::Arc<std::sync::atomic::AtomicU64>) -> Self {
-        Self {
-            slot,
-            released: false,
-        }
-    }
-
-    fn release(&mut self) {
-        if !self.released {
-            self.slot.store(0, std::sync::atomic::Ordering::Release);
-            self.released = true;
-        }
-    }
-}
-
-impl Drop for RelaySlotGuard {
-    fn drop(&mut self) {
-        if !self.released {
-            // #2841 (codex review): reaching Drop without a prior explicit
-            // release() means an abnormal exit (panic / `?` / task
-            // cancellation) BEFORE the turn recorded its relayed offset /
-            // advanced confirmed-end — so the delivery outcome of any in-flight
-            // Discord send is UNKNOWN. Freeing the slot prevents a permanent
-            // channel wedge, but a replacement watcher MAY then re-emit the same
-            // range (a bounded duplicate window). This is strictly better than a
-            // permanent wedge; the (channel, turn, byte-range) delivery lease
-            // (P1) closes the window by recording delivery BEFORE the slot
-            // frees. Surface it so the window is measurable until the lease lands.
-            tracing::warn!(
-                target: "agentdesk::relay_flight_recorder",
-                "relay emission slot freed via Drop on abnormal exit (in-flight send outcome unknown); a replacement watcher may re-emit the same range — resolved by the delivery lease"
-            );
-        }
-        self.release();
-    }
-}
-
-async fn wait_for_session_bound_relay_delivery_ack(
-    target: Option<&SessionBoundRelayAckTarget>,
-    timeout: std::time::Duration,
-) -> SessionBoundRelayAckOutcome {
-    if target.is_none() {
-        return SessionBoundRelayAckOutcome::MissingTarget;
-    }
-    let deadline = tokio::time::Instant::now() + timeout;
-    loop {
-        if let Some(outcome) = session_bound_relay_ack_snapshot_outcome(target) {
-            return outcome;
-        }
-        let now = tokio::time::Instant::now();
-        if now >= deadline {
-            return SessionBoundRelayAckOutcome::TimedOut;
-        }
-        tokio::time::sleep(std::time::Duration::from_millis(25).min(deadline - now)).await;
-    }
-}
-
-fn terminal_event_consumed_offset(current_offset: u64, unprocessed_tail: &str) -> u64 {
-    current_offset.saturating_sub(unprocessed_tail.len() as u64)
-}
-
-/// #3041 P1-3 (Part a, B1 — frame-carried commit fence): build the
-/// `TerminalCommitFence` to ride on the RESULT-bearing chunk's frame, or `None`
-/// when this chunk is not the terminal one / has no real consumed range / has no
-/// pinned turn identity to gate the sink's advance.
-///
-/// The fence carries the watcher's AUTHORITATIVE consumed-terminal `end`
-/// (`terminal_event_consumed_offset(current_offset, all_data)` == the watcher's own
-/// lease `end`) plus the PINNED turn identity (`user_msg_id` + `started_at`, #3141
-/// pinned-id semantics — from the turn-start inflight snapshot, filtered to THIS
-/// session). The sink advances `confirmed_end_offset` to `end` on a CONFIRMED
-/// delivery ONLY when this identity still matches the channel's current inflight
-/// (delayed-old-frame / wrong-turn protection). We DO NOT gate on `sink_can_own`:
-/// the fence is inert unless the sink confirms a delivery AND the identity matches,
-/// so carrying it on every real terminal chunk is safe.
-fn watcher_terminal_commit_fence(
-    found_result: bool,
-    turn_data_start_offset: u64,
-    consumed_end: u64,
-    pinned_identity: Option<&crate::services::discord::inflight::InflightTurnIdentity>,
-    tmux_session_name: &str,
-) -> Option<crate::services::cluster::stream_relay::TerminalCommitFence> {
-    if !found_result || consumed_end <= turn_data_start_offset {
-        return None;
-    }
-    let identity = pinned_identity?;
-    // Only pin an identity for THIS tmux session (the panel-identity snapshot is
-    // already filtered to it, but guard defensively so a cross-session snapshot
-    // can never seed a wrong-turn fence).
-    if identity.tmux_session_name.as_deref() != Some(tmux_session_name) {
-        return None;
-    }
-    // #3041 P1-3 (codex P1-3 issue 2 R4): a fence MUST carry a real
-    // `turn_start_offset`. The sink's identity gate is STRICT on this field (no
-    // None fallback) so two consecutive `user_msg_id == 0` (TUI-direct) turns in
-    // the same one-second `started_at` cannot collide. If the would-be terminal
-    // turn's start offset is unknown, DO NOT emit a fence: returning None forwards
-    // a non-terminal frame instead, and the watcher reconciliation's SendFull then
-    // safely delivers this turn (no black-hole, no weakly-gated advance).
-    let turn_start_offset = identity.turn_start_offset?;
-    Some(
-        crate::services::cluster::stream_relay::TerminalCommitFence {
-            consumed_end,
-            turn_user_msg_id: identity.user_msg_id,
-            turn_started_at: identity.started_at.clone(),
-            // A fence ALWAYS carries a real `turn_start_offset` (guaranteed above)
-            // so the sink's identity gate can disambiguate two consecutive
-            // `user_msg_id == 0` turns started in the same second.
-            turn_start_offset: Some(turn_start_offset),
-        },
-    )
 }
 
 /// Resolve the provider session selector to durably persist at turn end.
@@ -8224,14 +7486,14 @@ pub(in crate::services::discord) async fn tmux_output_watcher_with_restore(
 #[cfg(test)]
 mod tests {
     use super::{
-        FreshIdleFinalizeDecision, RelaySlotGuard, SessionBoundRelayAckOutcome,
-        TuiCompletionGateOutcome, Utf8ChunkDecoder,
-        adopt_watcher_terminal_message_ids_from_inflight, build_watcher_streaming_edit_text,
+        FreshIdleFinalizeDecision, SessionBoundRelayAckOutcome, TuiCompletionGateOutcome,
+        Utf8ChunkDecoder, adopt_watcher_terminal_message_ids_from_inflight,
+        build_watcher_streaming_edit_text,
         discard_restored_response_seed_before_no_inflight_terminal_relay,
         discard_watcher_pending_buffer_after_suppressed_turn,
         legacy_wrapper_prompt_candidates_from_pane, mark_watcher_terminal_delivery_committed,
         reacquire_watcher_inflight_for_active_stream, resolve_persistable_provider_session_id,
-        should_probe_tmux_liveness, terminal_event_consumed_offset, terminal_relay_decision,
+        should_probe_tmux_liveness, terminal_relay_decision,
         watcher_batch_contains_assistant_event, watcher_batch_contains_relayable_response,
         watcher_direct_terminal_should_commit_session_idle,
         watcher_fallback_edit_failure_can_delete_original_placeholder,
@@ -8275,12 +7537,6 @@ mod tests {
         }
     }
 
-    #[test]
-    fn terminal_event_consumed_offset_excludes_buffered_tail() {
-        assert_eq!(terminal_event_consumed_offset(128, "next-turn\n"), 118);
-        assert_eq!(terminal_event_consumed_offset(8, "longer-than-offset"), 0);
-    }
-
     // #3095: a freshly observed TUI session id always wins so the DB tracks the
     // newest selector.
     #[test]
@@ -8321,23 +7577,6 @@ mod tests {
             Some("cached-sid".to_string()),
             "blank observed must fall through to the usable cached selector"
         );
-    }
-
-    #[test]
-    fn relay_slot_guard_releases_on_drop() {
-        use std::sync::Arc;
-        use std::sync::atomic::{AtomicU64, Ordering};
-
-        // Simulate a watcher acquiring the slot (CAS 0 -> non-zero token).
-        let slot = Arc::new(AtomicU64::new(0));
-        slot.store(42, Ordering::Release);
-        {
-            let _guard = RelaySlotGuard::new(slot.clone());
-            assert_eq!(slot.load(Ordering::Acquire), 42, "slot held inside scope");
-        }
-        // #2840: dropping without an explicit release (panic / `?` / abort) must
-        // still free the slot so a replacement watcher is not wedged.
-        assert_eq!(slot.load(Ordering::Acquire), 0, "Drop released the slot");
     }
 
     #[test]
@@ -10123,32 +9362,6 @@ mod tests {
     }
 
     #[test]
-    fn relay_slot_guard_release_is_idempotent_and_does_not_clobber_reacquire() {
-        use std::sync::Arc;
-        use std::sync::atomic::{AtomicU64, Ordering};
-
-        let slot = Arc::new(AtomicU64::new(7));
-        let mut guard = RelaySlotGuard::new(slot.clone());
-        guard.release();
-        assert_eq!(
-            slot.load(Ordering::Acquire),
-            0,
-            "explicit release frees slot"
-        );
-
-        // After the explicit release, another watcher may legitimately acquire
-        // the slot. The first guard's trailing Drop must NOT reset that token to
-        // 0 — the idempotent `released` flag guarantees it.
-        slot.store(99, Ordering::Release);
-        drop(guard);
-        assert_eq!(
-            slot.load(Ordering::Acquire),
-            99,
-            "Drop after explicit release must not clobber a re-acquired slot"
-        );
-    }
-
-    #[test]
     fn bridge_suppressed_turn_discards_pending_buffer_before_direct_input() {
         let mut all_data = "{\"type\":\"assistant\",\"message\":\"old\"}\n".to_string();
         let mut all_data_start_offset = 10;
@@ -11012,284 +10225,6 @@ TUI-E2E-marker ssh-direct
                 !cell.renew(holder, turn, extended.saturating_add(1)),
                 "a renew on a Committed lease (a late tick after commit) is a no-op"
             );
-        }
-    }
-
-    /// #3151: the deterministic decision seam for the in-flight sink-delivery
-    /// marker gate (`watcher_terminal_resend_action_gated`). Table-drives the gate
-    /// over every lease-snapshot variant and asserts the reclaim side-effect flag.
-    /// The decision fn is PURE (no cell mutation) so the side effect is testable in
-    /// isolation; the integration tests below exercise the actual `reclaim_if_expired`.
-    mod inflight_sink_marker_gate {
-        use super::super::{
-            WatcherTerminalResendAction, watcher_terminal_resend_action,
-            watcher_terminal_resend_action_gated,
-        };
-        use crate::services::discord::turn_finalizer::TurnKey;
-        use crate::services::discord::{
-            DeliveryLeaseCell, LeaseHolder, LeaseOutcome, LeaseSnapshot, lease_now_ms,
-        };
-        use serenity::model::id::ChannelId;
-
-        const START: u64 = 100;
-        const END: u64 = 200;
-        // `committed < end` so the underlying reconciliation would choose SendFull.
-        const COMMITTED_BELOW_END: u64 = 100;
-        const NOW: u64 = 50_000;
-
-        fn turn() -> TurnKey {
-            TurnKey::new(ChannelId::new(7201), 9, 0)
-        }
-
-        /// Unleased → behaves EXACTLY as the ungated reconciliation (SendFull when
-        /// committed<end), no reclaim.
-        #[test]
-        fn unleased_defers_to_reconciliation() {
-            let (action, reclaim) = watcher_terminal_resend_action_gated(
-                &LeaseSnapshot::Unleased,
-                COMMITTED_BELOW_END,
-                START,
-                END,
-                NOW,
-            );
-            assert_eq!(action, WatcherTerminalResendAction::SendFull);
-            assert!(!reclaim);
-            // ... and committed>=end on an Unleased cell still Skips (unchanged).
-            let (skip, reclaim2) = watcher_terminal_resend_action_gated(
-                &LeaseSnapshot::Unleased,
-                END,
-                START,
-                END,
-                NOW,
-            );
-            assert_eq!(skip, WatcherTerminalResendAction::SkipAlreadyCommitted);
-            assert!(!reclaim2);
-        }
-
-        /// Leased{Sink, FRESH} (now < deadline) → WaitInFlight, no reclaim. This is
-        /// the slow-sink-in-flight case: the watcher must NOT re-send this pass.
-        #[test]
-        fn leased_sink_fresh_waits_in_flight() {
-            let snap = LeaseSnapshot::Leased {
-                holder: LeaseHolder::Sink,
-                turn: turn(),
-                deadline_ms: NOW + 5_000, // fresh: deadline strictly in the future
-                start: START,
-                end: END,
-            };
-            let (action, reclaim) =
-                watcher_terminal_resend_action_gated(&snap, COMMITTED_BELOW_END, START, END, NOW);
-            assert_eq!(action, WatcherTerminalResendAction::WaitInFlight);
-            assert!(!reclaim, "a fresh sink lease must NOT be reclaimed");
-        }
-
-        /// Leased{Sink, EXPIRED} (now >= deadline) → reclaim flag set AND SendFull
-        /// (committed<end). This is the dead-sink no-black-hole arm.
-        #[test]
-        fn leased_sink_expired_reclaims_and_sends_full() {
-            let snap = LeaseSnapshot::Leased {
-                holder: LeaseHolder::Sink,
-                turn: turn(),
-                deadline_ms: NOW, // expired: now >= deadline
-                start: START,
-                end: END,
-            };
-            let (action, reclaim) =
-                watcher_terminal_resend_action_gated(&snap, COMMITTED_BELOW_END, START, END, NOW);
-            assert_eq!(action, WatcherTerminalResendAction::SendFull);
-            assert!(
-                reclaim,
-                "an expired sink lease MUST be reclaimed (no black-hole)"
-            );
-        }
-
-        /// #3159 BUG 1: Committed{Sink, Delivered} with committed >= end (a real
-        /// delivered commit advances the offset BEFORE committing) → Skip, no reclaim.
-        /// This is the no-duplicate invariant: a genuinely-delivered range is never
-        /// re-sent.
-        #[test]
-        fn committed_sink_delivered_covered_skips() {
-            let snap = LeaseSnapshot::Committed {
-                holder: LeaseHolder::Sink,
-                turn: turn(),
-                start: START,
-                end: END,
-                outcome: LeaseOutcome::Delivered,
-            };
-            // committed >= end (END): the advance ran before commit.
-            let (action, reclaim) =
-                watcher_terminal_resend_action_gated(&snap, END, START, END, NOW);
-            assert_eq!(action, WatcherTerminalResendAction::SkipAlreadyCommitted);
-            assert!(!reclaim);
-        }
-
-        /// #3159 BUG 1 (no black-hole): Committed{Sink, NotDelivered} — the identity
-        /// gate REFUSED the advance, so committed stayed < end. The gate now routes
-        /// through committed-offset reconciliation → SendFull (re-send, not Skip).
-        /// Previously this blind-Skipped → under-delivery / black-hole.
-        #[test]
-        fn committed_sink_not_delivered_sends_full() {
-            let snap = LeaseSnapshot::Committed {
-                holder: LeaseHolder::Sink,
-                turn: turn(),
-                start: START,
-                end: END,
-                outcome: LeaseOutcome::NotDelivered,
-            };
-            let (action, reclaim) =
-                watcher_terminal_resend_action_gated(&snap, COMMITTED_BELOW_END, START, END, NOW);
-            assert_eq!(
-                action,
-                WatcherTerminalResendAction::SendFull,
-                "a NotDelivered commit (committed<end) must re-send, not Skip"
-            );
-            assert!(!reclaim);
-        }
-
-        /// #3159 BUG 1 belt-and-suspenders: even a Delivered-labelled commit with
-        /// committed < end (which the fixed producer no longer emits, since Delivered
-        /// is committed only after a real advance) re-sends — the committed offset is
-        /// the SOLE delivered-test, not the outcome label. No black-hole regardless.
-        #[test]
-        fn committed_sink_below_end_sends_full_regardless_of_label() {
-            let snap = LeaseSnapshot::Committed {
-                holder: LeaseHolder::Sink,
-                turn: turn(),
-                start: START,
-                end: END,
-                outcome: LeaseOutcome::Delivered,
-            };
-            let (action, reclaim) =
-                watcher_terminal_resend_action_gated(&snap, COMMITTED_BELOW_END, START, END, NOW);
-            assert_eq!(action, WatcherTerminalResendAction::SendFull);
-            assert!(!reclaim);
-        }
-
-        /// Leased by a WATCHER (non-Sink) holder → the #3151 gate does NOT interpose;
-        /// it defers to the existing reconciliation (the B2 path is untouched).
-        #[test]
-        fn leased_by_watcher_defers_to_reconciliation() {
-            let snap = LeaseSnapshot::Leased {
-                holder: LeaseHolder::Watcher { instance_id: 1 },
-                turn: turn(),
-                deadline_ms: NOW + 5_000,
-                start: START,
-                end: END,
-            };
-            // committed<end → SendFull (NOT WaitInFlight: only a Sink lease waits).
-            let (action, reclaim) =
-                watcher_terminal_resend_action_gated(&snap, COMMITTED_BELOW_END, START, END, NOW);
-            assert_eq!(action, WatcherTerminalResendAction::SendFull);
-            assert!(!reclaim);
-            // committed>=end on a watcher-held lease still Skips.
-            let (skip, _) = watcher_terminal_resend_action_gated(&snap, END, START, END, NOW);
-            assert_eq!(skip, WatcherTerminalResendAction::SkipAlreadyCommitted);
-        }
-
-        /// committed>=end with a Bridge holder → Skip (the range is delivered),
-        /// matching the ungated path.
-        #[test]
-        fn committed_covered_skips_for_non_sink() {
-            let snap = LeaseSnapshot::Leased {
-                holder: LeaseHolder::Bridge,
-                turn: turn(),
-                deadline_ms: NOW + 1,
-                start: START,
-                end: END,
-            };
-            let (action, _) = watcher_terminal_resend_action_gated(&snap, END, START, END, NOW);
-            assert_eq!(action, WatcherTerminalResendAction::SkipAlreadyCommitted);
-            // Sanity: the gated decision equals the ungated reconciliation here.
-            assert_eq!(action, watcher_terminal_resend_action(END, START, END));
-        }
-
-        /// (b) Integration: a DEAD/STALE sink marker on a real cell is reclaimed by
-        /// the gate's `reclaim_if_expired` side effect, then the watcher re-acquires
-        /// and SendFulls — NO black-hole. Drives the actual cell, not just the flag.
-        #[test]
-        fn dead_sink_marker_reclaimed_then_resent_no_blackhole() {
-            let ch = ChannelId::new(7202);
-            let cell = DeliveryLeaseCell::new(ch);
-            let sink_turn = TurnKey::new(ch, 9, 0);
-            let now = lease_now_ms();
-            let deadline = now.saturating_add(10);
-            // Sink set the marker then "died" (no heartbeat renews it).
-            assert!(cell.try_acquire(sink_turn, LeaseHolder::Sink, START, END, deadline));
-
-            // The gate, observed at a time PAST the deadline, decides reclaim+SendFull.
-            let past = deadline.saturating_add(1);
-            let snap = cell.read();
-            let (action, reclaim) =
-                watcher_terminal_resend_action_gated(&snap, COMMITTED_BELOW_END, START, END, past);
-            assert_eq!(action, WatcherTerminalResendAction::SendFull);
-            assert!(reclaim);
-
-            // The caller performs the reclaim → the dead marker clears → a
-            // replacement watcher re-acquires and re-delivers (no black-hole).
-            assert!(cell.reclaim_if_expired(past));
-            assert!(matches!(cell.read(), LeaseSnapshot::Unleased));
-            let watcher_turn = TurnKey::new(ch, 9, 0);
-            assert!(
-                cell.try_acquire(
-                    watcher_turn,
-                    LeaseHolder::Watcher { instance_id: 1 },
-                    START,
-                    END,
-                    past.saturating_add(10_000),
-                ),
-                "the reclaimed cell is re-acquirable by the watcher (no black-hole)"
-            );
-        }
-
-        /// (c) reclaim-races-with-late-sink-success cannot corrupt the lease: after a
-        /// dead sink's marker is reclaimed and the watcher re-acquires, the zombie
-        /// sink's late `commit`/`release` (full-identity-gated) NO-OP against the
-        /// watcher's lease — no wrong-holder advance, no stolen release.
-        #[test]
-        fn reclaim_then_late_sink_commit_cannot_corrupt_lease() {
-            let ch = ChannelId::new(7203);
-            let cell = DeliveryLeaseCell::new(ch);
-            let sink_turn = TurnKey::new(ch, 9, 0);
-            let now = lease_now_ms();
-            let deadline = now.saturating_add(10);
-            assert!(cell.try_acquire(sink_turn, LeaseHolder::Sink, START, END, deadline));
-
-            // Watcher reclaims the expired sink marker and re-acquires the SAME range.
-            let past = deadline.saturating_add(1);
-            assert!(cell.reclaim_if_expired(past));
-            let watcher_holder = LeaseHolder::Watcher { instance_id: 1 };
-            assert!(cell.try_acquire(
-                sink_turn,
-                watcher_holder,
-                START,
-                END,
-                past.saturating_add(10_000),
-            ));
-
-            // The zombie sink's LATE commit/release target Sink+sink_turn; the cell
-            // is now held by the Watcher → both no-op (false). No corruption.
-            assert!(
-                !cell.commit(
-                    LeaseHolder::Sink,
-                    sink_turn,
-                    START,
-                    END,
-                    LeaseOutcome::Delivered
-                ),
-                "a late sink commit must NOT act on the watcher's lease"
-            );
-            assert!(
-                !cell.release(LeaseHolder::Sink, sink_turn, START, END),
-                "a late sink release must NOT free the watcher's lease"
-            );
-            // The watcher's lease is intact and committable by its true holder.
-            assert!(cell.commit(
-                watcher_holder,
-                sink_turn,
-                START,
-                END,
-                LeaseOutcome::Delivered
-            ));
         }
     }
